feat(substate)!: widen the substate version to u64
A substate version counts one write per transaction that touches it, so its
ceiling is a throughput ceiling on a single contended substate rather than a
calendar one. u32 exhausts in ~50 days at 1,000 writes/s on one substate and
~1.4 years at 100/s - both rates a hot vault or AMM pool can reach - where u64
puts the ceiling out of reach entirely.

Overflow was also unsafe rather than a clean reject: every increment was a plain
`+ 1`, and the version is the low bytes of `SubstateAddress`, so a wrapped
version collides with the v0 address and reuses a JMT key for a substate that is
already down.

Consensus-breaking. The version is chained into `hash_substate` through Borsh,
which is fixed-width, so every substate value hash changes; `SubstateAddress`
grows 36 -> 40 bytes, so every JMT key changes with it; and the version inside a
transaction's input requirements is part of the signed preimage. Networks must
reset rather than upgrade in place.

Version routing is untouched: the version is a pure suffix of the address and
`to_shard` reads only its first byte.

Also:
- Pin the version fields to a TS `number` so the generated bindings are
  unchanged; ts-rs maps a bare u64 to `bigint`, and a version stays far below
  2^53.
- Widen the SQLite version columns in the wallet and indexer stores. The wallet's
  `vaults` is a foreign key target, so its column is replaced in place rather
  than through a table rebuild, which inside a migration's transaction would
  repoint every referencing table at the temporary name.
- Raise the `ProposedBlockChangeSet` memory guardrail to 15MB; the wider version
  fields in `SubstateChange` and `SubstateLock` move the worst case from
  13.4MB to 14.3MB.

+++ applications/tari_indexer/src/storage_sqlite/models/substate_cache.rs
@@ -8,7 +8,7 @@ use tari_indexer_lib::substate_cache::caches_nonexistence;
 pub(crate) struct SubstateCacheRow {
     #[allow(dead_code)]
     pub substate_id: String,
-    pub version: Option<i32>,
+    pub version: Option<i64>,
     pub verified: bool,
     pub substate_result: Vec<u8>,
     pub cached_at: i64,
@@ -23,9 +23,9 @@ pub(crate) struct SubstateCacheRow {
 #[derive(Debug, Clone)]
 pub struct SubstateCacheInvalidation {
     substate_id: SubstateId,
-    retires_up_to: Option<u32>,
+    retires_up_to: Option<u64>,
     retires_nonexistence: bool,
-    observed_version: u32,
+    observed_version: u64,
     spent: bool,
 }
 
@@ -37,7 +37,7 @@ impl SubstateCacheInvalidation {
     /// retraction of a cached nonexistence. Where nothing caches that, it carries nothing, and is
     /// not one of these at all: emitting it would put a journal row on the sync path for every
     /// created-once substate in the stream and buy nothing with it.
-    pub fn created(substate_id: &SubstateId, version: u32) -> Option<Self> {
+    pub fn created(substate_id: &SubstateId, version: u64) -> Option<Self> {
         let retires_up_to = version.checked_sub(1);
         let retires_nonexistence = caches_nonexistence(substate_id);
         if retires_up_to.is_none() && !retires_nonexistence {
@@ -56,7 +56,7 @@ impl SubstateCacheInvalidation {
     ///
     /// A destroy leaves a cached nonexistence alone. `DoesNotExist` says the substate has no live
     /// version, which a destroy makes more true rather than less.
-    pub fn destroyed(substate_id: SubstateId, version: u32) -> Self {
+    pub fn destroyed(substate_id: SubstateId, version: u64) -> Self {
         Self {
             substate_id,
             retires_up_to: Some(version),
@@ -72,7 +72,7 @@ impl SubstateCacheInvalidation {
 
     /// The version the stream showed the substate at: created at it, or destroyed at it. A head
     /// below this is one the substate has already been watched past, whoever offers it.
-    pub fn observed_version(&self) -> u32 {
+    pub fn observed_version(&self) -> u64 {
         self.observed_version
     }
 
@@ -85,7 +85,7 @@ impl SubstateCacheInvalidation {
     }
 
     /// The highest cached head version this retires, if any.
-    pub fn retires_up_to(&self) -> Option<u32> {
+    pub fn retires_up_to(&self) -> Option<u64> {
         self.retires_up_to
     }
